# services/application-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH 2026 Food Traceability - Application Service",
    "description": "SIH 2026 Developer 3 Application / Backend Microservice - Handles Auth/RBAC, Public APIs, Workflow Orchestration, QR Resolution, Feedback Accountability, Risk Propagator & Recall Actions.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create Product",
        "description": "Create and register a new food product definition.",
        "operationId": "create_product_api_v1_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/products/{product_id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get Product",
        "description": "Retrieve product definition by product ID.",
        "operationId": "get_product_api_v1_products__product_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches": {
      "post": {
        "tags": [
          "Batches"
        ],
        "summary": "Create Batch",
        "description": "Create and register a new traceable product batch.",
        "operationId": "create_batch_api_v1_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/batches/{batch_id}/validate": {
      "post": {
        "tags": [
          "Batches"
        ],
        "summary": "Validate Batch",
        "description": "Validate a batch state through the authorized validation workflow (P0 Milestone).",
        "operationId": "validate_batch_api_v1_batches__batch_id__validate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}/transfer": {
      "post": {
        "tags": [
          "Batches"
        ],
        "summary": "Transfer Custody",
        "description": "Transfer batch custody to another supply-chain organization.",
        "operationId": "transfer_custody_api_v1_batches__batch_id__transfer_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustodyTransferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}": {
      "get": {
        "tags": [
          "Batches"
        ],
        "summary": "Get Batch",
        "description": "Retrieve batch details and current lifecycle state.",
        "operationId": "get_batch_api_v1_batches__batch_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units": {
      "post": {
        "tags": [
          "Units"
        ],
        "summary": "Create Unit",
        "description": "Create individual sellable/traceable physical unit derived from a batch.",
        "operationId": "create_unit_api_v1_units_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/units/{unit_id}": {
      "get": {
        "tags": [
          "Units"
        ],
        "summary": "Get Unit",
        "description": "Retrieve details of a specific unit by unit ID.",
        "operationId": "get_unit_api_v1_units__unit_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchCreate": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id",
            "example": "prd-oj-001"
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "example": 500.0
          },
          "unit_of_measure": {
            "type": "string",
            "title": "Unit Of Measure",
            "default": "KG",
            "example": "KG"
          },
          "parent_batch_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Batch Ids",
            "example": []
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationData"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "title": "BatchCreate"
      },
      "BatchResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "producer_org_id": {
            "type": "string",
            "title": "Producer Org Id"
          },
          "current_custodian_org_id": {
            "type": "string",
            "title": "Current Custodian Org Id"
          },
          "lifecycle_state": {
            "type": "string",
            "title": "Lifecycle State"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "unit_of_measure": {
            "type": "string",
            "title": "Unit Of Measure"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "blockchain_tx_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockchain Tx Id"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "product_id",
          "producer_org_id",
          "current_custodian_org_id",
          "lifecycle_state",
          "quantity",
          "unit_of_measure",
          "created_at"
        ],
        "title": "BatchResponse"
      },
      "BatchValidateRequest": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "example": "Quality inspection passed"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationData"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "BatchValidateRequest"
      },
...
      "CustodyTransferRequest": {
        "properties": {
          "to_org_id": {
            "type": "string",
            "title": "To Org Id",
            "example": "org-transporter-fastlogistics"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "example": "Dispatching shipment"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationData"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "to_org_id"
        ],
        "title": "CustodyTransferRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationData": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "example": 19.2183
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "example": 72.9781
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name",
            "example": "Lokmanya Nagar, Thane, Maharashtra"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "LocationData"
      },
...
      "ProductCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Fresh Premium Orange Juice"
          },
          "sku": {
            "type": "string",
            "title": "Sku",
            "example": "SKU-ORJ-500ML"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "BEVERAGE",
            "example": "BEVERAGE"
          },
          "specifications": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specifications",
            "example": {
              "origin": "California",
              "volume": "500ml"
            }
          }
        },
        "type": "object",
        "required": [
          "name",
          "sku"
        ],
        "title": "ProductCreate"
      },
...
      "ProductResponse": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "producer_org_id": {
            "type": "string",
            "title": "Producer Org Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "blockchain_tx_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockchain Tx Id"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "name",
          "sku",
          "category",
          "producer_org_id",
          "created_at"
        ],
        "title": "ProductResponse"
      },
...
      "UnitCreate": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id",
            "example": "batch-raw-101"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number",
            "example": "SN-9988112233"
          }
        },
        "type": "object",
        "required": [
          "batch_id"
        ],
        "title": "UnitCreate"
      },
      "UnitResponse": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "serial_number": {
            "type": "string",
            "title": "Serial Number"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "blockchain_tx_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockchain Tx Id"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "batch_id",
          "serial_number",
          "status",
          "created_at"
        ],
        "title": "UnitResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}